# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Game:
    properties:
      bet_amount:
        description: nanotons
        type: integer
      complete_tx_hash:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      fee_receiver_address:
        type: string
      game_id:
        type: integer
      highest_bid_allowed:
        type: integer
      init_tx_hash:
        type: string
      join_tx_hash:
        type: string
      joined_at:
        type: string
      lowest_bid_allowed:
        type: integer
      min_referrer_payout_value:
        type: integer
      payout_amount:
        type: integer
      player_one_address:
        type: string
      player_one_choice:
        description: 1=CLOSED, 2=HEADS, 3=TAILS
        type: integer
      player_one_referrer:
        type: string
      player_two_address:
        type: string
      player_two_choice:
        type: integer
      player_two_referrer:
        type: string
      protocol_version:
        type: integer
      referrer_fee_numerator:
        type: integer
      reveal_tx_hash:
        type: string
      revealed_at:
        type: string
      service_fee_numerator:
        type: integer
      status:
        description: 0-4 (see constants above)
        type: integer
      waiting_timeout_seconds:
        type: integer
      winner_address:
        type: string
    type: object
...
  rest.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  rest.GameListResponse:
    properties:
      games:
        items:
          $ref: '#/definitions/rest.GameWithReservation'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  rest.GameWithReservation:
    properties:
      game:
        $ref: '#/definitions/entity.Game'
      reservation:
        $ref: '#/definitions/rest.Reservation'
    type: object
...
  rest.Reservation:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      game_id:
        type: integer
      status:
        type: string
      wallet_address:
        type: string
    type: object
...
host: localhost:8090
info:
  contact:
    name: POD Game Team
  description: Backend service for TON blockchain gambling game with real-time WebSocket
    updates
  license:
    name: BUSL-1.1
  title: POD Game Backend API
  version: 1.0.0
paths:
...
  /api/v1/games:
    get:
      consumes:
      - application/json
      description: Retrieve list of games filtered by status with pagination support.
        Game payloads may include remediated factory config/protocol fields projected
        from on-chain events; reservation data is advisory and on-chain status remains
        authoritative.
      parameters:
      - default: 1
        description: Filter by game status (0-4)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        in: query
        name: status
        type: integer
      - default: 20
        description: Maximum number of results
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of games
          schema:
            $ref: '#/definitions/rest.GameListResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get available games
      tags:
      - games
  /api/v1/games/{gameId}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information for a specific game by ID, including
        projected remediated factory config/protocol fields when available. Backend
        read models are indexed views of on-chain state.
      parameters:
      - description: Game ID
        in: path
        minimum: 1
        name: gameId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Game details
          schema:
            $ref: '#/definitions/entity.Game'
        "400":
          description: Invalid game ID
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Game not found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get game details
      tags:
      - games
...
schemes:
- http
- https
securityDefinitions:
  TelegramAuth:
    description: Telegram Mini App authentication via initData string
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"